Fix totalFrames calculation for non-LAME MP3s

+++ Audio/Properties/MPEGPropertiesReader.m
@@ -337,7 +337,7 @@
 			}
 			
 			// Just estimate the number of frames based on the file's size
-			unsigned totalFrames = (double)frame.header.samplerate * ((stat.st_size - id3_length) / (frame.header.bitrate / 8.0));
+			unsigned totalFrames = frame.header.samplerate * ((stat.st_size - id3_length) / (frame.header.bitrate / 8.0));
 
 			[propertiesDictionary setValue:[NSNumber numberWithUnsignedLong:totalFrames] forKey:PropertiesTotalFramesKey];
